pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "directors-cut"
version = "0.1.0"
description = "Autonomous Video Editing Agent MCP Server"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "moviepy>=1.0.3",
    "librosa>=0.10.0",
    "scenedetect>=0.6.2",
    "openai>=1.0.0",
    "yt-dlp>=2023.10.13",
    "youtube-transcript-api>=0.6.0",
    "google-generativeai>=0.3.0",
    "python-dotenv>=1.0.0",
    "numpy>=1.24.0",
    "gradio[mcp]>=4.0.0",
    "elevenlabs>=0.2.26",
    "requests>=2.31.0",
    "whisperx>=3.1.1",
    "torch>=2.0.0",
]

[project.scripts]
director = "src.server:main"
